feat(windows): stub unsupported compute drivers
