+++ crates/openshell-driver-docker/Cargo.toml
@@ -12,14 +12,15 @@ repository.workspace = true
 
 [dependencies]
 openshell-core = { path = "../openshell-core", default-features = false }
+serde = { workspace = true }
 
+[target.'cfg(not(target_os = "windows"))'.dependencies]
 tokio = { workspace = true }
 tonic = { workspace = true }
 futures = { workspace = true }
 tokio-stream = { workspace = true }
 tracing = { workspace = true }
 bytes = { workspace = true }
-serde = { workspace = true }
 serde_json = { workspace = true }
 prost-types = { workspace = true }
 bollard = { version = "0.20" }
