+++ crates/openshell-driver-kubernetes/Cargo.toml
@@ -17,7 +17,9 @@ path = "src/main.rs"
 [dependencies]
 openshell-core = { path = "../openshell-core", default-features = false }
 openshell-policy = { path = "../openshell-policy" }
+serde = { workspace = true }
 
+[target.'cfg(not(target_os = "windows"))'.dependencies]
 tokio = { workspace = true }
 tonic = { workspace = true, features = ["transport"] }
 prost = { workspace = true }
@@ -27,7 +29,6 @@ tokio-stream = { workspace = true }
 kube = { workspace = true }
 kube-runtime = { workspace = true }
 k8s-openapi = { workspace = true }
-serde = { workspace = true }
 serde_json = { workspace = true }
 clap = { workspace = true }
 tracing = { workspace = true }
@@ -37,6 +38,7 @@ miette = { workspace = true }
 
 [dev-dependencies]
 temp-env = "0.3"
+serde_json = { workspace = true }
 
 [lints]
 workspace = true
