+++ crates/openshell-driver-kubernetes/src/lib.rs
@@ -2,7 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod config;
+
+#[cfg(not(target_os = "windows"))]
 pub mod driver;
+#[cfg(not(target_os = "windows"))]
 pub mod grpc;
 
 pub use config::{
@@ -10,5 +13,8 @@ pub use config::{
     DEFAULT_WORKSPACE_STORAGE_SIZE, KubernetesComputeConfig, KubernetesSidecarConfig,
     SupervisorSideloadMethod, SupervisorTopology,
 };
+
+#[cfg(not(target_os = "windows"))]
 pub use driver::{KubernetesComputeDriver, KubernetesDriverError};
+#[cfg(not(target_os = "windows"))]
 pub use grpc::ComputeDriverService;
