+++ crates/openshell-driver-podman/Cargo.toml
@@ -14,7 +14,10 @@ repository.workspace = true
 name = "openshell-driver-podman"
 path = "src/main.rs"
 
-[dependencies]
+[target.'cfg(target_os = "windows")'.dependencies]
+serde = { workspace = true }
+
+[target.'cfg(not(target_os = "windows"))'.dependencies]
 openshell-core = { path = "../openshell-core", default-features = false }
 
 tokio = { workspace = true }
