+++ crates/openshell-driver-podman/src/lib.rs
@@ -1,15 +1,11 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-pub(crate) mod client;
-pub mod config;
-pub(crate) mod container;
-pub mod driver;
-pub mod grpc;
-#[cfg(test)]
-pub(crate) mod test_utils;
-pub(crate) mod watcher;
+#[cfg(target_os = "windows")]
+mod windows;
 
-pub use config::PodmanComputeConfig;
-pub use driver::PodmanComputeDriver;
-pub use grpc::ComputeDriverService;
+#[cfg(target_os = "windows")]
+pub use windows::PodmanComputeConfig;
+
+#[cfg(not(target_os = "windows"))]
+include!("lib_unix.rs");
