+++ crates/openshell-driver-podman/src/lib_unix.rs
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) mod client;
+pub mod config;
+pub(crate) mod container;
+pub mod driver;
+pub mod grpc;
+#[cfg(test)]
+pub(crate) mod test_utils;
+pub(crate) mod watcher;
+
+pub use config::PodmanComputeConfig;
+pub use driver::PodmanComputeDriver;
+pub use grpc::ComputeDriverService;
