+++ crates/openshell-driver-vm/Cargo.toml
@@ -18,7 +18,7 @@ path = "src/lib.rs"
 name = "openshell-driver-vm"
 path = "src/main.rs"
 
-[dependencies]
+[target.'cfg(not(target_os = "windows"))'.dependencies]
 openshell-core = { path = "../openshell-core", default-features = false }
 openshell-otel = { path = "../openshell-otel" }
 openshell-policy = { path = "../openshell-policy" }
