+++ crates/openshell-driver-vm/src/lib.rs
@@ -1,24 +1,10 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-pub mod driver;
-mod embedded_runtime;
-mod ffi;
-pub mod gpu;
-pub mod lifecycle;
-mod nft_ruleset;
-pub mod otel_tracing;
-pub mod procguard;
-mod rootfs;
-mod runtime;
+#[cfg(target_os = "windows")]
+pub fn windows_unsupported() -> &'static str {
+    "openshell-driver-vm is unsupported on Windows"
+}
 
-pub use driver::{VmDriver, VmDriverConfig};
-pub use lifecycle::{
-    BackendFeature, ExtensionCapabilities, ExtensionDescriptor, GuestInitDropin, LaunchAbortReason,
-    LaunchPlan, LifecycleError, LifecycleExtension, LifecycleExtensionRegistry, LifecycleResult,
-    RestoreContext,
-};
-pub use runtime::{
-    VM_RUNTIME_DIR_ENV, VmBackend, VmLaunchConfig, cleanup_stale_tap_interfaces,
-    configured_runtime_dir, run_vm,
-};
+#[cfg(not(target_os = "windows"))]
+include!("lib_unix.rs");
