+++ crates/openshell-server/src/compute/mod.rs
@@ -42,11 +42,21 @@ use openshell_core::proto::{
     SandboxTemplate, ServiceEndpoint, SshSession,
 };
 use openshell_core::{ObjectLabels, ObjectWorkspace};
+#[cfg(not(target_os = "windows"))]
 use openshell_driver_docker::DockerComputeDriver;
+#[cfg(target_os = "windows")]
+use openshell_driver_kubernetes::KubernetesComputeConfig;
+#[cfg(not(target_os = "windows"))]
 use openshell_driver_kubernetes::{
-    ComputeDriverService as KubernetesDriverService, KubernetesComputeDriver,
+    ComputeDriverService as KubernetesDriverService, KubernetesComputeConfig,
+    KubernetesComputeDriver,
 };
-use openshell_driver_podman::{ComputeDriverService as PodmanDriverService, PodmanComputeDriver};
+#[cfg(target_os = "windows")]
+use openshell_driver_podman::PodmanComputeConfig;
+#[cfg(not(target_os = "windows"))]
+use openshell_driver_podman::{
+    ComputeDriverService as PodmanDriverService, PodmanComputeConfig, PodmanComputeDriver,
+};
 use prost::Message;
 use std::collections::HashMap;
 use std::fmt;
@@ -268,6 +278,7 @@ trait ShutdownCleanup: Send + Sync {
 }
 
 #[tonic::async_trait]
+#[cfg(not(target_os = "windows"))]
 impl ShutdownCleanup for DockerComputeDriver {
     async fn cleanup_on_shutdown(&self) -> Result<(), String> {
         let stopped = self
@@ -293,6 +304,7 @@ trait StartupResume: Send + Sync {
 }
 
 #[tonic::async_trait]
+#[cfg(not(target_os = "windows"))]
 impl StartupResume for DockerComputeDriver {
     async fn resume_sandbox(&self, sandbox_id: &str, sandbox_name: &str) -> Result<bool, String> {
         Self::resume_sandbox(self, sandbox_id, sandbox_name)
@@ -310,6 +322,11 @@ const ORPHAN_GRACE_PERIOD: Duration = Duration::from_secs(300);
 // Re-export the shared error type under the name used by this module.
 pub use openshell_core::ComputeDriverError as ComputeError;
 
+#[cfg(target_os = "windows")]
+fn unsupported_compute_driver_error(driver: &str) -> ComputeError {
+    ComputeError::Message(format!("{driver} compute driver is unsupported on Windows"))
+}
+
 #[derive(Debug)]
 pub struct ManagedDriverProcess {
     child: std::sync::Mutex<Option<tokio::process::Child>>,
@@ -317,6 +334,7 @@ pub struct ManagedDriverProcess {
 }
 
 impl ManagedDriverProcess {
+    #[cfg(unix)]
     pub(crate) fn new(child: tokio::process::Child, socket_path: PathBuf) -> Self {
         Self {
             child: std::sync::Mutex::new(Some(child)),
@@ -710,6 +728,7 @@ impl ComputeRuntime {
         self.delete_gates.entry_count()
     }
 
+    #[cfg(not(target_os = "windows"))]
     pub async fn new_docker(
         config: openshell_core::Config,
         docker_config: DockerComputeConfig,
@@ -742,6 +761,20 @@ impl ComputeRuntime {
         .await
     }
 
+    #[cfg(target_os = "windows")]
+    pub async fn new_docker(
+        _config: openshell_core::Config,
+        _docker_config: DockerComputeConfig,
+        _store: Arc<Store>,
+        _sandbox_index: SandboxIndex,
+        _sandbox_watch_bus: SandboxWatchBus,
+        _tracing_log_bus: TracingLogBus,
+        _supervisor_sessions: Arc<SupervisorSessionRegistry>,
+    ) -> Result<Self, ComputeError> {
+        Err(unsupported_compute_driver_error("Docker"))
+    }
+
+    #[cfg(not(target_os = "windows"))]
     pub async fn new_kubernetes(
         config: KubernetesComputeConfig,
         store: Arc<Store>,
@@ -769,6 +802,18 @@ impl ComputeRuntime {
         .await
     }
 
+    #[cfg(target_os = "windows")]
+    pub async fn new_kubernetes(
+        _config: KubernetesComputeConfig,
+        _store: Arc<Store>,
+        _sandbox_index: SandboxIndex,
+        _sandbox_watch_bus: SandboxWatchBus,
+        _tracing_log_bus: TracingLogBus,
+        _supervisor_sessions: Arc<SupervisorSessionRegistry>,
+    ) -> Result<Self, ComputeError> {
+        Err(unsupported_compute_driver_error("Kubernetes"))
+    }
+
     pub(crate) async fn new_remote_driver(
         endpoint: AcquiredRemoteDriverEndpoint,
         store: Arc<Store>,
@@ -793,6 +838,7 @@ impl ComputeRuntime {
         .await
     }
 
+    #[cfg(not(target_os = "windows"))]
     pub async fn new_podman(
         config: PodmanComputeConfig,
         store: Arc<Store>,
@@ -818,6 +864,18 @@ impl ComputeRuntime {
             supervisor_sessions,
         )
         .await
+    }
+
+    #[cfg(target_os = "windows")]
+    pub async fn new_podman(
+        _config: PodmanComputeConfig,
+        _store: Arc<Store>,
+        _sandbox_index: SandboxIndex,
+        _sandbox_watch_bus: SandboxWatchBus,
+        _tracing_log_bus: TracingLogBus,
+        _supervisor_sessions: Arc<SupervisorSessionRegistry>,
+    ) -> Result<Self, ComputeError> {
+        Err(unsupported_compute_driver_error("Podman"))
     }
 
     #[must_use]
@@ -7245,5 +7303,17 @@ mod tests {
             .unwrap()
             .unwrap();
         assert_eq!(stored.object_workspace(), "alpha");
+    }
+
+    #[cfg(target_os = "windows")]
+    #[test]
+    fn windows_compute_driver_stubs_report_unsupported() {
+        for driver in ["Docker", "Kubernetes", "Podman"] {
+            let message = unsupported_compute_driver_error(driver).to_string();
+            assert!(
+                message.contains("unsupported on Windows"),
+                "{driver} stub message should be explicit, got: {message}"
+            );
+        }
     }
 }
